Move ButtonDarkMode component into ui-components
